Fix multple compiler warnings

+++ filter/string.cpp
@@ -137,7 +137,7 @@ std::string implode (const std::vector <std::string>& values, std::string delimi
 // Example input {1, 2, 3, 4}, max three last bits, the output will be [12, 3, 4}.
 void implode_from_beginning_remain_with_max_n_bits (std::vector<std::string>& input, const int n, const std::string& joiner)
 {
-  while (input.size() > n) {
+  while (static_cast<int>(input.size()) > n) {
     input[1].insert(0, input.at(0) + joiner);
     input.erase(input.cbegin());
   }
